requirements.txt
discord.py==2.0.0a4489+gf1ec0919
pendulum==2.1.2
psycopg2-binary==2.9.3
pytz==2022.1
